+++ src/frontend/wdisp/windisp.c
@@ -436,7 +436,7 @@ LRESULT CALLBACK PlotWindowProc( HWND hwnd,
             angle = RAD_TO_DEG * atan2( fy0, fx0 );
             fprintf(stdout, "r0 = %g, a0 = %g\n",
             sqrt( fx0*fx0 + fy0*fy0 ),
-            (angle>0)?angle:(double) 360+angle);
+            (angle>0)?angle:360.0+angle);
          }
       } else  {    
          /* need to print info about two points */
